requirements.txt
Django~=3.2.7
django-shortuuidfield~=0.1.3
requests~=2.26.0
Js2Py~=0.71
beautifulsoup4~=4.10.0
twilio